Adding full support for deep-links in Android Kotlin.
