+++ samples/kotlin-android-app/src/main/java/com/segment/analytics/next/MainActivity.kt
@@ -1,11 +1,13 @@
 package com.segment.analytics.next
 
+import android.content.Intent
 import android.os.Bundle
 import android.widget.Button
 import androidx.appcompat.app.AppCompatActivity
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.FragmentManager
 import androidx.fragment.app.FragmentTransaction
+import com.segment.analytics.kotlin.android.addDeepLinkOpen
 import com.segment.analytics.kotlin.core.Analytics
 import com.segment.analytics.kotlin.core.EventType
 import com.segment.analytics.kotlin.core.platform.plugins.logger.log
@@ -51,5 +53,19 @@ class MainActivity : AppCompatActivity() {
         // replace the FrameLayout with new Fragment
         fragmentTransaction.replace(R.id.frameLayout, fragment)
         fragmentTransaction.commit() // save the changes
+    }
+
+    override fun onNewIntent(intent: Intent?) {
+        super.onNewIntent(intent)
+
+        // Add a manual deep-link opened event.
+        // This is necessary when your Activity has a android:launchMode of
+        // 'singleInstance', 'singleInstancePerTask', 'singleTop', or any other mode
+        // that will re-use an existing Activity instead of creating a new instance.
+        // The Analytics SDK automatically identifies when you app is started from a deep link
+        // if the Activity is created, but not if it is re-used. Therefore we have to add this
+        // code to manually capture the Deep Link info.
+
+        Analytics.addDeepLinkOpen(analytics, "deep-link", intent)
     }
 }
